+++ src/_Commands/Discord/analytics.command.js
@@ -100,7 +100,7 @@ module.exports = {
 				borders: true,
 				grid: true,
 				minimumY: 0
-			});
+			}).image.getBufferAs('membercount.png');
 		}
 	}
 };