+++ src/_Commands/Moderation/moderation.command.js
@@ -56,14 +56,14 @@ module.exports = {
 		}
 	},
 	'archive': {
-		aliases: ['move'],
+		aliases: ['move','mirror'],
 		category: 'Moderation',
 		//title: 'Archive',
 		info: 'Move messages in the current channel to an archive channel. Use flags to specify the kinds of messages to target: `-cmds`, `-bot`, `-media`, `-text`, and `-pinned`.',
 		parameters: ['count|...messages', '[channel]'],
 		flags: ['b|bot','c|cmds','t|text','m|media','p|pinned'],
 		permissions: 'privileged',
-		fn({client, server, channel, args, flags}) {
+		fn({client, server, channel, cmd, args, flags}) {
 			if (typeof(args[0]) === 'number') {
 				return Moderation.archive(client, server, channel, args[0], flags);
 			} else {
@@ -75,7 +75,8 @@ module.exports = {
 				return client.moveMessages({
 					from: channel.id,
 					to: channelID,
-					messages: args
+					messages: args,
+					keepOriginalMessages: (cmd.toLowerCase() == 'mirror')
 				});
 			}
 		},